lab-5-docker-compose/requirements.txt
Flask==2.3.0
redis==4.5.1
psycopg2-binary==2.9.3